Feat: Dev Pengembalian penjualan

+++ src/app/components/pengeluaran/listpengeluaran/(table)/deleteButton.tsx
@@ -2,6 +2,7 @@
 
 import { deleteData } from "@/app/(public)/(main)/pengeluaran/listpengeluaran/action"
 import useStore from "@/app/(public)/(main)/pengeluaran/listpengeluaran/store"
+import { passwordInputAlert } from "@/utils/alert/swal"
 import { useState } from "react"
 import Swal from "sweetalert2"
 
@@ -9,12 +10,17 @@ const DeleteButton = ({ id }: {
     id: string
 }) => {
     const [isLoading, setIsLoading] = useState<boolean>(false)
-    const remove = useStore(state => state.remove)
+    const reset = useStore(state => state.reset)
     const handleClick = async () => {
         try{
-            setIsLoading(true)
-            await deleteData([id])
-            setIsLoading(false)
+            const confirmPassword = await passwordInputAlert()
+
+            if(confirmPassword){
+                setIsLoading(true)
+                await deleteData([id])
+                setIsLoading(false)
+                reset()
+            }
         }catch{
             Swal.fire({
                 icon: 'error',
